Finish refactoring components
