+++ src/components/spatial/MouseCoordinates/index.css
