+++ src/components/spatial/MouseCoordinates/index.tsx
@@ -5,9 +5,9 @@ import { Feature, FeatureCollection, LineString, MultiPoint, Point, Polygon } fr
 import * as turf from "@turf/turf";
 import nearestPoint from '@turf/nearest-point';
 import './index.css';
-import { useAppContext } from '../../context/AppContext';
-import { useAppSelector } from '../../app/hooks';
-import { formatCoordinate } from '../../helpers/formatCoordinate';
+import { useAppContext } from '../../../context/AppContext';
+import { useAppSelector } from '../../../app/hooks';
+import { formatCoordinate } from '../../../helpers/formatCoordinate';
 
 const bearingToAzimuth = (bearing: number) => {
   return (bearing + 360) % 360;
