+++ src/components/spatial/Track/index.tsx
@@ -3,8 +3,8 @@ import { LeafletMouseEvent } from 'leaflet';
 import { Polyline, CircleMarker, Tooltip } from 'react-leaflet';
 import { format } from "date-fns";
 import { useMemo } from "react";
-import { useAppContext } from "../../context/AppContext";
-import { CoordInstance, filterTrack } from "../../helpers/filterTrack";
+import { useAppContext } from "../../../context/AppContext";
+import { CoordInstance, filterTrack } from "../../../helpers/filterTrack";
 
 export interface TrackFeatureProps {
   feature: Feature
